splitting scan command
