+++ test/azure/commands/scan.spec.ts
@@ -22,9 +22,15 @@ describe('Scan Integration Tests', function () {
   const group2VNetId = `/subscriptions/${subscriptionId}/resourceGroups/${resourceGroup2}/providers/Microsoft.Network/virtualNetworks/azatestvnet2`;
   test
     .stdout()
-    .command(['scan', '--scope', subscriptionId, '-f', '../test/rules.json'])
+    .command([
+      'scan:rg',
+      '--subscription',
+      subscriptionId,
+      '-f',
+      '../test/rules.json',
+    ])
     .it(
-      'runs scan --scope [subscriptionId] -f ../test/rules.json',
+      'runs scan --subscription [subscriptionId] -f ../test/rules.json',
       async ({stdout}) => {
         const scanner = new Scanner();
         await scanner.loadRulesFromFile('../test/rules.json');
@@ -37,7 +43,7 @@ describe('Scan Integration Tests', function () {
   test
     .stdout()
     .command([
-      'scan',
+      'scan:rg',
       '-s',
       subscriptionId,
       '-g',
@@ -55,7 +61,7 @@ describe('Scan Integration Tests', function () {
   test
     .stdout()
     .command([
-      'scan',
+      'scan:rg',
       '-s',
       subscriptionId,
       '-g',
@@ -75,7 +81,7 @@ describe('Scan Integration Tests', function () {
   test
     .stderr()
     .command([
-      'scan',
+      'scan:rg',
       '-s',
       subscriptionId,
       '-g',
@@ -94,7 +100,7 @@ describe('Scan Integration Tests', function () {
     );
   test
     .stdout()
-    .command(['scan', '-s', subscriptionId, '-g', resourceGroup])
+    .command(['scan:rg', '-s', subscriptionId, '-g', resourceGroup])
     .it(
       'should find storage accounts with a private endpoint configured but the public endpoint is still enabled',
       async ({stdout}) => {
