+++ test/unit/commands/scan.spec.ts
@@ -7,17 +7,27 @@ describe('Scan Unit Tests', function () {
     .stdout()
     .command(['scan'])
     .exit(2)
-    .it('exits with error code 2 when running scan without a flag');
+    .it('exits with error code 2 when running scan without a rule type');
   test
     .stdout()
     .stderr()
-    .command(['scan', '--scope'])
+    .command(['scan:rg', '--subscription'])
     .exit(2)
-    .it('exits with error code 2 when running scan --scope without a value');
+    .it(
+      'exits with error code 2 when running scan --subscription without a value'
+    );
   test
     .stdout()
     .stderr()
-    .command(['scan', '-s', 'subId1', '-s', 'subId2', '-g', 'resourceGroup'])
+    .command(['scan:arm', '--subscription'])
+    .exit(2)
+    .it(
+      'exits with error code 2 when running scan --subscription without a value'
+    );
+  test
+    .stdout()
+    .stderr()
+    .command(['scan:rg', '-s', 'subId1', '-s', 'subId2', '-g', 'resourceGroup'])
     .exit(2)
     .it(
       'exits with error code 2 when running providing multiple subscriptions and 1 or more resource groups'
